django-backend/requirements.txt
django
djangorestframework
wheel
coreapi
